Move LH5 write tests to dedicated file
